Add basic impl of REINFORCE
The current impl is for the 24 game. Ops are chosen via a neural net,
but arguments are chosen randomly. This makes the agent behave very
noisily, so it is hard to determine if the REINFORCE algorithm is
behaving properly. We should implement neural net based argument
and then study what REINFORCE is doing more closely.

I also started using the MLflow library to log experiments. See the
policy_gradient.py file for an example of how it is used, or ask Tony
via DM.